rename account route and add no cache
